requirements.txt
Django==5.1.5
django-add-default-value==0.10.0
django-autoslug==1.9.9
django-brotli==0.2.1
django-cache-machine==1.2.0
django-cors-headers==4.3.1
django-countries==7.6.1
django-cprofile-middleware==1.0.5
django-crispy-forms==2.1
django-csp==3.8
django-enumfield==3.1
django-environ==0.11.2
django-extensions==3.2.3
django-guardian==2.4.0
django-ical==1.9.2
django-ipware==7.0.1
django-jsonify==0.3.0
django-json-widget==2.0.1
django-ltree-2==0.1.8
django-modeltranslation==0.19.12
django-oauth-toolkit==2.4.0
django-phonenumber-field==7.3.0
django-pivot==1.10.0
django-print-sql==2018.3.6
django-ratelimit==4.1.0
django-rq==3.0.0
django-sql-utils==0.7.0
django-static-compress==2.0.0
django-static-fontawesome==6.5.2.0
django-super-deduper==0.1.4
django-tastypie==0.14.7
django-user-agents==0.4.0
django-webpush==0.3.6
git+https://github.com/appfluence/django-tastypie-swagger@763f8f027ec5cc6e20c33812cd8dec5b32404583
git+https://github.com/shtalinberg/django-el-pagination@c0885a0625c1698efba1076ae8cc9764763bcb10
pytest-django==4.8.0
phonenumbers==8.13.38
psycopg2-binary==2.9.9
arrow==1.3.0
tqdm==4.66.4
requests==2.32.3
pytimeparse==1.1.8
humanfriendly==10.0
humanize==4.9.0
python-telegram-bot==13.15
lxml==5.2.2
defusedxml==0.7.1
PyYAML==6.0.1
biplist==1.0.3
ipdb==0.13.13
google-api-python-client==1.11.0
oauth2client==4.1.3
python-memcached==1.62
django-debug-toolbar==4.4.2
yapf==0.40.2
Unidecode==1.3.8
first==2.0.2
setuptools
jsonfield==3.1.0
geopy==2.4.1
cssutils==2.11.1
unicodecsv==0.14.1
ratelimiter==1.2.0
docutils==0.21.2
image==1.5.33
Pillow==10.3.0
olefile==0.47
pytesseract==0.3.10
python-dateutil==2.9.0
scikit-image==0.23.2
string-color==1.2.3
channels==4.1.0
flake8==7.0.0
isort==5.13.2
emoji-country-flag==1.3.2
coloredlogs==15.0.1
channels-redis==4.2.0
fire==0.6.0
strictfire==0.4.1
lazy-load==0.8.3
blackboxprotobuf==1.0.1
traceback-with-variables==2.0.4
flatten-dict==0.4.2
filelock==3.14.0
tailslide==0.1.1
feedgen==1.0.0
pycountry==24.6.1
multiset==3.1.0
twisted[tls,http2]==24.7.0
uWSGI==2.0.26
supervisor==4.2.5
daphne==4.1.2
tabulate==0.9.0
sentry-sdk==2.8.0
chardet==5.2.0
python-magic==0.4.27
prettytable==3.10.0
watchdog==4.0.1
pymemcache==4.0.0
crispy-bootstrap3==2024.1
rich==13.7.1
numpy==1.26.4
scipy==1.13.1
numba==0.59.1
ipwhois==1.2.0
fontawesomefree==6.5.1
brotli==1.0.9
zstandard==0.23.0
geoip2==4.8.0
xgboost==2.1.1
scikit-learn==1.5.1
pandas==2.2.2
beautifulsoup4==4.12.3
markdown==3.7
